meme_refiner/requirements.txt
google-adk>=0.3.0
python-dotenv>=1.0.0
mcp>=1.0.0
rich>=13.0.0
duckduckgo-search
beautifulsoup4
requests
litellm
cohere
asyncpg==0.29.0
psycopg2-binary==2.9.9
greenlet
sqlalchemy==2.0.25
fastapi
uvicorn
ddgs